+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.4.6] - 2024-11-05
+
 ### Added
 
 - `--http-header` CLI argument ([#238](https://github.com/stac-utils/stac-asset/pull/238))
@@ -215,7 +217,8 @@ Used to be <https://github.com/gadomski/stac-asset>.
 
 Initial release.
 
-[unreleased]: https://github.com/stac-utils/stac-asset/compare/v0.4.5...HEAD
+[unreleased]: https://github.com/stac-utils/stac-asset/compare/v0.4.6...HEAD
+[0.4.6]: https://github.com/stac-utils/stac-asset/compare/v0.4.5...v0.4.6
 [0.4.5]: https://github.com/stac-utils/stac-asset/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/stac-utils/stac-asset/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/stac-utils/stac-asset/compare/v0.4.2...v0.4.3